+++ contao/dca/tl_module.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Contao Etracker Bundle
+ * etracker Plugin for Contao
  *
  * (c) Xenbyte, Stefan Brauner <info@xenbyte.com>
  *
@@ -12,10 +12,6 @@
 declare(strict_types=1);
 
 use Contao\CoreBundle\DataContainer\PaletteManipulator;
-use Xenbyte\ContaoEtracker\Controller\FrontendModule\EtrackerOptoutController;
-
-// Palette for etracker Optout Module
-$GLOBALS['TL_DCA']['tl_module']['palettes'][EtrackerOptoutController::TYPE] = '{title_legend},name,headline,type;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
 
 // Palettes for Search Module
 $GLOBALS['TL_DCA']['tl_module']['palettes']['__selector__'][] = 'etrackerSearchCampaignEnable';
@@ -45,7 +41,7 @@ $GLOBALS['TL_DCA']['tl_module']['fields']['etrackerSearchMedOnsite'] = [
         'tl_class'  => 'w50',
         'maxlength' => 50,
     ],
-    'sql'       => 'varchar(50) NULL default NULL',
+    'sql'       => 'varchar(50) NULL',
 ];
 
 $GLOBALS['TL_DCA']['tl_module']['fields']['etrackerSearchCmpOnsiteResults'] = [
@@ -54,7 +50,7 @@ $GLOBALS['TL_DCA']['tl_module']['fields']['etrackerSearchCmpOnsiteResults'] = [
         'tl_class'  => 'w50',
         'maxlength' => 50,
     ],
-    'sql'       => 'varchar(50) NULL default \'mit Ergebnis\'',
+    'sql'       => 'varchar(50) NULL',
 ];
 
 $GLOBALS['TL_DCA']['tl_module']['fields']['etrackerSearchCmpOnsiteNoresults'] = [
@@ -63,5 +59,5 @@ $GLOBALS['TL_DCA']['tl_module']['fields']['etrackerSearchCmpOnsiteNoresults'] = 
         'tl_class'  => 'w50',
         'maxlength' => 50,
     ],
-    'sql'       => 'varchar(50) NULL default \'ohne Ergebnis\'',
+    'sql'       => 'varchar(50) NULL',
 ];