+++ contao/languages/en/tl_form_field.php
@@ -15,5 +15,5 @@ declare(strict_types=1);
 $GLOBALS['TL_LANG']['tl_form_field']['etracker_legend'] = 'etracker';
 
 // fields
-$GLOBALS['TL_LANG']['tl_form_field']['etrackerFieldname'] = ['abweichender Name', 'Name des Formularfeldes (sonst Feldbezeichnung oder Feldname)'];
-$GLOBALS['TL_LANG']['tl_form_field']['etrackerFormSection'] = ['Name der "Sektion"', '(Standard: "Standard")'];
+$GLOBALS['TL_LANG']['tl_form_field']['etrackerFieldname'] = ['deviating name', 'Name of the form field (otherwise field description or field name)'];
+$GLOBALS['TL_LANG']['tl_form_field']['etrackerFormSection'] = ['Name of the "section"', '(Default: "Standard")'];