+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/lang/de/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Added
+
+- Search result campaign tracking
+- Opt out module
+
+### Changed
+
+- Moved templates
+- Removed FrontendModule with ContentElement
+
+### Fixed
+
+- Using _etrackerOnReady for events in form tracking
+
 ## [0.3.0]
 
 ### Added